(Bug 229, P3: It is possible to assign the same keyboard shortcut to 2 commands)
Ed Musgrove's patch to fix cases where the bug currently occurs. 

This does not completely fix the bug, because users can still set modifier keys in the wrong order so they would be two separate entries in CommandManager::mCommandKeyHash, but KeyEventToKeyString() will put them in the expected order, so the version in the expected order will always be the one CommandManager::HandleKey() uses.

git-svn-id: https://audacity.googlecode.com/svn/audacity-src/trunk@10700 47890b92-0858-11df-a26f-8b716316a5bc