Fix Bug 448 (P2) - WAV export corrupt if imported file has metadata after data (edit)
Applies patch from libsndfile 1.2.25 on the git repo.

src/strings.c : Fix handling of SF_STR_SOFTWARE that resulted in a segfault.

FYI, The commit has the following properties
commit 386af45517724a1b10a790899c8c4ec2c4b161af
Author: Erik de Castro Lopo <erikd@mega-nerd.com>
Date:   Fri Mar 25 19:12:24 2011 +1100

  src/strings.c : Fix handling of SF_STR_SOFTWARE that resulted in a segfault.

Also, there is another patch that may relate to bug 448 that deals with the same code:
Since I am not sure if that is needed or not, I will just note it for now, and this commit will only contain one patch:
commit 7194d455dbf9b4927057eacd443800d24ec40a19
Author: Erik de Castro Lopo <erikd@mega-nerd.com>
Date:   Tue Jun 28 18:22:12 2011 +1000

    src/strings.c : Clean up and refactor storage of SF_STR_SOFTWARE.

Because these are already in the next version of libsndfile I don't see a reason to note the patches

git-svn-id: https://audacity.googlecode.com/svn/audacity-src/trunk@11750 47890b92-0858-11df-a26f-8b716316a5bc