Mouse wheel caused hang.
mScrollRemainder was not getting initialised and so the first time TimeTextCtrl::OnMouse got called, 'steps' got set to a random number (very large and -ve here) and Adjust got called with -2147483648, -1.  So we need to initialise mScrollRemainder.

This has been wrong for a while, from what I see.  Previous version used Decrease and Increase which were immune to the problem.

'Adjust' decrements steps, whether it is positive or negative to start with.  It is designed to have 'steps' strictly positive, so make sure of that.

Make sure OnMouse uses Adjust correctly.

git-svn-id: https://audacity.googlecode.com/svn/audacity-src/trunk@12771 47890b92-0858-11df-a26f-8b716316a5bc