Add buffer delay compensation and settings dialog
The buffer delay compensation resolves the issue where an effect
may add a delay to the output.  This would cause the output to shift
and the final samples to get chopped off.

The settings (click Settings on any VST dialog) dialog allows the user
to enable/disable the BDC and to set their desired buffer size (within
reason) to help improve performance.

In addition it fixes a bug in the plugin scan list were you could click
OK twice and have two scans going, eventually asserting.



git-svn-id: https://audacity.googlecode.com/svn/audacity-src/trunk@13102 47890b92-0858-11df-a26f-8b716316a5bc